REF: Move persistence and credentials into utils.cloud package
