+++ octue/resources/communication/google_pub_sub/service.py
@@ -7,9 +7,9 @@ from google.cloud import pubsub_v1
 
 from octue import exceptions
 from octue.mixins import CoolNameable
-from octue.resources.communication.credentials import GCPCredentialsManager
 from octue.resources.communication.google_pub_sub import Subscription, Topic
 from octue.resources.manifest import Manifest
+from octue.utils.cloud.credentials import GCPCredentialsManager
 
 
 logger = logging.getLogger(__name__)
